TP_Seminar/03_twitter_bot/requirements.txt
tweepy
Pillow
python-resize-image